+++ sdk-ts/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.4] - 2026-06-20
+
 ### Changed
 
 - Make the public `ai-hist` command Rust-default for the user-facing CLI